refactor: create team-affiliations service
