+++ src/components/tables/team-affiliations.tsx
@@ -2,8 +2,8 @@ import { TFunction, useTranslation } from 'next-i18next'
 import { ReactNode } from 'react'
 
 import { Table } from '@/components/tables/common/table'
+import { TeamAffiliationsSortBy } from '@/modules/team-affiliations.ts/types'
 import { ICommonTableProps } from '@/types/tables'
-import { TeamAffiliationsSortBy } from '@/types/team-affiliations'
 
 interface ITeamAffiliationsTableProps extends ICommonTableProps {
   children: ReactNode
