+++ src/pages/players/[slug].tsx
@@ -8,11 +8,11 @@ import { PageHeading } from '@/components/page-heading/page-heading'
 import { TeamAffiliationsTableRow } from '@/components/tables/rows/team-affiliations-row'
 import { TeamAffiliationsTable } from '@/components/tables/team-affiliations'
 import { withSessionSsr } from '@/lib/session'
-import { useTeamAffiliations } from '@/lib/team-affiliations'
 import { PlayerDto } from '@/modules/players/types'
+import { useTeamAffiliations } from '@/modules/team-affiliations.ts/hooks'
 import { getPlayerBySlug } from '@/services/api/methods/players'
 import { ApiError } from '@/types/common'
-import { TeamAffiliationsSortBy } from '@/types/team-affiliations'
+import { TeamAffiliationsSortBy } from '@/modules/team-affiliations.ts/types'
 import { useTable } from '@/utils/hooks/use-table'
 import { redirectToLogin } from '@/utils/redirect-to-login'
