+++ src/services/api/methods/team-affiliations.ts
@@ -1,0 +1,14 @@
+import {
+  FindAllTeamAffiliationsParams,
+  TeamAffiliationDto,
+} from '@/modules/team-affiliations.ts/types'
+import { getPaginatedData } from '@/services/api/methods/helpers'
+import { TModuleName } from '@/services/api/modules'
+
+const moduleName: TModuleName = 'team-affiliations'
+
+export const getTeamAffiliations = (params: FindAllTeamAffiliationsParams) =>
+  getPaginatedData<FindAllTeamAffiliationsParams, TeamAffiliationDto>(
+    params,
+    moduleName,
+  )
